requirements.txt
langchain==0.1.16
langchain-community==0.0.32
langchain-openai==0.0.8
openai==1.14.3
python-dotenv==1.0.0
streamlit==1.22.0
tiktoken==0.5.1
unstructured==0.9.2
chromadb==0.4.24
